+ Add Account Info     Just total money

+++ presentation/src/main/java/com/tinyapps/presentation/features/transactions/model/Wallet.kt
@@ -1,3 +1,3 @@
 package com.tinyapps.presentation.features.transactions.model
 
-data class Wallet(val name : String, val balance: Long)
+data class Wallet(val name : String, val balance: Double)
